moved common for display and merge
