+++ modules/derivation/src/main/scala-2/tofu/common/derived/DisplayDerivationImpl.scala
@@ -6,18 +6,12 @@ import magnolia1.{CaseClass, Magnolia, SealedTrait}
 import tofu.common.Display
 import tofu.magnolia.compat
 
-/** Derivation of [[Display]] typeclass for case classes and sealed traits
-  *
-  * @note
-  *   Derived [[Display]] instances will indent nested structures if those are supposed to be on newline. You can see
-  *   examples in the tests.
-  */
-object display extends Derivation[Display] {
+trait DisplayDerivationImpl extends Derivation[Display] {
 
   private type Typeclass[T] = Display[T]
 
   def join[T](ctx: CaseClass[Typeclass, T]): Display[T]    = (cfg: Display.Config, a: T) => {
-    import cfg.*
+    import cfg._
 
     val nestedIndent = indent + indent
