+++ modules/derivation/src/main/scala-3/tofu/common/derived/DisplayDerivationImpl.scala
@@ -7,13 +7,7 @@ import tofu.magnolia.compat
 import scala.deriving.Mirror
 
 
-/** Derivation of [[Display]] typeclass for case classes and sealed traits
-  *
-  * @note
-  *   Derived [[Display]] instances will indent nested structures if those are supposed to be on newline. You can see
-  *   examples in the tests.
-  */
-object display extends AutoDerivation[Display] {
+trait DisplayDerivationImpl extends AutoDerivation[Display] {
 
   private type Typeclass[T] = Display[T]
