+++ modules/derivation/src/main/scala-3/tofu/common/derived/package.scala
@@ -2,8 +2,6 @@ package tofu.common
 
 import scala.deriving.Mirror
 
-
 package object derived {
-      extension (x: Display.type) inline def derived[A](using Mirror.Of[A]): Display[A] = display.derived[A]
-
+  extension (x: Display.type) inline def derived[A](using Mirror.Of[A]): Display[A] = display.derived[A]
 }
