+++ modules/derivation/src/main/scala/tofu/common/derived/display.scala
@@ -1,0 +1,12 @@
+package tofu.common.derived
+
+import tofu.common.Display
+
+
+/** Derivation of [[Display]] typeclass for case classes and sealed traits
+ *
+ * @note
+ *   Derived [[Display]] instances will indent nested structures if those are supposed to be on newline. You can see
+ *   examples in the tests.
+ */
+object display extends DisplayDerivationImpl
